app/requirements.txt
streamlit==1.29.0
pillow==10.1.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
python-dotenv==1.0.0